Fix bad ouptut from nc-config
I was getting "yes" output from some of the options which were not enabled. This was due to two reasons:

If CMake set the variable to "OFF", then the -z $var test would fail and $var would be set to "yes" even though it should be "no"
If CMake set the variable to empty string (has_var=""), then the -z $has_var test would also fail and set it to yes instead of no.
With these changes, it looks like I am getting the correct output from nc-config consistent with my configuration options.